+++ src/racing-plus/configDescription.ts
@@ -476,10 +476,19 @@ export const QUALITY_OF_LIFE_CHANGES_1: ConfigDescriptions = [
     ],
   ],
   [
-    "combinedDualityDoors", // 498
+    "leadPencilChargeBar", // 444
     [
       ModConfigMenuOptionType.BOOLEAN,
       "058",
+      "Show Holy Mantle on the hearts UI",
+      "Show the presence of a Holy Mantle shield on the hearts UI.",
+    ],
+  ],
+  [
+    "combinedDualityDoors", // 498
+    [
+      ModConfigMenuOptionType.BOOLEAN,
+      "059",
       "Duality revamp",
       "Make Duality combine the Devil Room door and the Angel Room door together.",
     ],
@@ -488,18 +497,9 @@ export const QUALITY_OF_LIFE_CHANGES_1: ConfigDescriptions = [
     "removeFortuneCookieBanners", // 557
     [
       ModConfigMenuOptionType.BOOLEAN,
-      "059",
+      "060",
       "Remove Fortune Cookie banners",
       "Remove Fortune Cookie banners when don't get any pickups.",
-    ],
-  ],
-  [
-    "showDreamCatcherItem", // 566
-    [
-      ModConfigMenuOptionType.BOOLEAN,
-      "060",
-      "Show the Dream Catcher item",
-      "If you have Dream Catcher, draw the Treasure Room item while in the starting room of the floor.",
     ],
   ],
 ];
@@ -507,10 +507,19 @@ export const QUALITY_OF_LIFE_CHANGES_1: ConfigDescriptions = [
 // 061-06X
 export const QUALITY_OF_LIFE_CHANGES_2: ConfigDescriptions = [
   [
-    "fastLuna", // 589
+    "showDreamCatcherItem", // 566
     [
       ModConfigMenuOptionType.BOOLEAN,
       "061",
+      "Show the Dream Catcher item",
+      "If you have Dream Catcher, draw the Treasure Room item while in the starting room of the floor.",
+    ],
+  ],
+  [
+    "fastLuna", // 589
+    [
+      ModConfigMenuOptionType.BOOLEAN,
+      "062",
       "Fast Luna",
       "Make Moonlights from Luna able to be entered as soon as they spawn.",
     ],
@@ -519,7 +528,7 @@ export const QUALITY_OF_LIFE_CHANGES_2: ConfigDescriptions = [
     "fadeVasculitisTears", // 657
     [
       ModConfigMenuOptionType.BOOLEAN,
-      "062",
+      "063",
       "Fade Vasculitis tears",
       "Fade the tears that explode out of enemies when you have Vasculitis.",
     ],
@@ -528,7 +537,7 @@ export const QUALITY_OF_LIFE_CHANGES_2: ConfigDescriptions = [
     "removePerfectionVelocity", // 145
     [
       ModConfigMenuOptionType.BOOLEAN,
-      "063",
+      "064",
       "Remove Perfection velocity",
       "Make the Perfection trinket spawn in a stationary spot so that it won't go over a pit.",
     ],
@@ -537,7 +546,7 @@ export const QUALITY_OF_LIFE_CHANGES_2: ConfigDescriptions = [
     "removePerfectionOnEndFloors", // 145
     [
       ModConfigMenuOptionType.BOOLEAN,
-      "064",
+      "065",
       "Remove Perfection on end floors",
       "Prevent the Perfection trinket from spawning on the final floor of a run.",
     ],
@@ -546,7 +555,7 @@ export const QUALITY_OF_LIFE_CHANGES_2: ConfigDescriptions = [
     "automaticItemInsertion",
     [
       ModConfigMenuOptionType.BOOLEAN,
-      "065",
+      "066",
       "Automatically insert pickups",
       "When taking an item that drops pickups, automatically insert them into your inventory.",
     ],
@@ -555,7 +564,7 @@ export const QUALITY_OF_LIFE_CHANGES_2: ConfigDescriptions = [
     "chargePocketItemFirst",
     [
       ModConfigMenuOptionType.BOOLEAN,
-      "066",
+      "067",
       "Charge the Pocket item first",
       "Make batteries charge the pocket item first over the active item.",
     ],
@@ -564,7 +573,7 @@ export const QUALITY_OF_LIFE_CHANGES_2: ConfigDescriptions = [
     "showMaxFamiliars",
     [
       ModConfigMenuOptionType.BOOLEAN,
-      "067",
+      "068",
       "Show max familiars",
       "Show an icon on the UI when you have the maximum amount of familiars (i.e. 64).",
     ],
@@ -573,7 +582,7 @@ export const QUALITY_OF_LIFE_CHANGES_2: ConfigDescriptions = [
     "showPills",
     [
       ModConfigMenuOptionType.BOOLEAN,
-      "068",
+      "069",
       "Remember pills",
       "Hold the map button to see a list of identified pills for easy reference.",
     ],